+++ Libplanet/Store/LiteDBStore.cs
@@ -2,17 +2,18 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Runtime.Serialization.Formatters.Binary;
 using System.Security.Cryptography;
+using System.Text;
 using Libplanet.Action;
 using Libplanet.Blocks;
 using Libplanet.Serialization;
 using Libplanet.Tx;
 using LiteDB;
+using Serilog;
 
 namespace Libplanet.Store
 {
@@ -20,15 +21,19 @@ namespace Libplanet.Store
     /// <see cref="IStore"/> implementation using <a href="https://www.litedb.org/">LiteDB</a>.
     /// </summary>
     /// <seealso cref="IStore"/>
-    public class LiteDBStore : IStore, IDisposable
+    public class LiteDBStore : BaseStore, IDisposable
     {
         private const string TxIdPrefix = "tx/";
 
         private const string IndexColPrefix = "index_";
 
-        private const string StateRefIdPrefix = "stateref/";
+        private const string StateRefIdPrefix = "stateref_";
 
         private const string NonceIdPrefix = "nonce_";
+
+        private static HashAlgorithm namespaceHasher = MD5.Create();
+
+        private readonly ILogger _logger;
 
         private readonly LiteDatabase _db;
 
@@ -40,18 +45,28 @@ namespace Libplanet.Store
         /// Enables or disables double write check to ensure durability.
         /// </param>
         /// <param name="cacheSize">Max number of pages in the cache.</param>
-        public LiteDBStore(string path, bool journal = true, int cacheSize = 50000)
+        /// <param name="flush">Writes data direct to disk avoiding OS cache.  Turned on by default.
+        /// </param>
+        public LiteDBStore(
+            string path,
+            bool journal = true,
+            int cacheSize = 50000,
+            bool flush = true
+        )
         {
             if (path is null)
             {
                 throw new ArgumentNullException(nameof(path));
             }
 
+            _logger = Log.ForContext<LiteDBStore>();
+
             var connectionString = new ConnectionString
             {
                 Filename = path,
                 Journal = journal,
                 CacheSize = cacheSize,
+                Flush = flush,
             };
 
             if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX) &&
@@ -77,7 +92,7 @@ namespace Libplanet.Store
             _db.GetCollection<StagedTxIdDoc>("staged_txids");
 
         /// <inheritdoc/>
-        public IEnumerable<string> ListNamespaces()
+        public override IEnumerable<string> ListNamespaces()
         {
             return _db.GetCollectionNames()
                 .Where(name => name.StartsWith(IndexColPrefix))
@@ -85,25 +100,21 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public void DeleteNamespace(string @namespace)
+        public override void DeleteNamespace(string @namespace)
         {
             _db.DropCollection(IndexCollection(@namespace).Name);
             _db.DropCollection($"{NonceIdPrefix}{@namespace}");
-
-            foreach (LiteFileInfo file in _db.FileStorage.Find($"{StateRefIdPrefix}{@namespace}"))
-            {
-                _db.FileStorage.Delete(file.Id);
-            }
+            _db.DropCollection(StateRefId(@namespace));
         }
 
         /// <inheritdoc/>
-        public long CountIndex(string @namespace)
+        public override long CountIndex(string @namespace)
         {
             return IndexCollection(@namespace).Count();
         }
 
         /// <inheritdoc/>
-        public IEnumerable<HashDigest<SHA256>> IterateIndex(string @namespace)
+        public override IEnumerable<HashDigest<SHA256>> IterateIndex(string @namespace)
         {
             return IndexCollection(@namespace)
                 .FindAll()
@@ -111,7 +122,7 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public HashDigest<SHA256>? IndexBlockHash(string @namespace, long index)
+        public override HashDigest<SHA256>? IndexBlockHash(string @namespace, long index)
         {
             if (index < 0)
             {
@@ -127,14 +138,14 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public long AppendIndex(string @namespace, HashDigest<SHA256> hash)
+        public override long AppendIndex(string @namespace, HashDigest<SHA256> hash)
         {
             return IndexCollection(@namespace)
                        .Insert(new HashDoc { Hash = hash }) - 1;
         }
 
         /// <inheritdoc/>
-        public bool DeleteIndex(string @namespace, HashDigest<SHA256> hash)
+        public override bool DeleteIndex(string @namespace, HashDigest<SHA256> hash)
         {
             int deleted = IndexCollection(@namespace)
                 .Delete(i => i.Hash.Equals(hash));
@@ -142,40 +153,17 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public IEnumerable<Address> ListAddresses(string @namespace)
+        public override IEnumerable<Address> ListAddresses(string @namespace)
         {
-            var prefix = $"{StateRefIdPrefix}{@namespace}/";
-            foreach (LiteFileInfo fileInfo in _db.FileStorage.Find(prefix))
-            {
-                string fileId = fileInfo.Id;
-                int slashIndex = fileId.LastIndexOf('/');
-                if (slashIndex < 0)
-                {
-                    continue;
-                }
-
-                string addressHex = fileId.Substring(slashIndex + 1);
-                if (addressHex.Length < Address.Size * 2)
-                {
-                    continue;
-                }
-
-                Address address;
-                try
-                {
-                    address = new Address(addressHex);
-                }
-                catch (ArgumentException)
-                {
-                    continue;
-                }
-
-                yield return address;
-            }
+            string collId = StateRefId(@namespace);
+            return _db.GetCollection<StateRefDoc>(collId)
+                .Find(Query.All("Address", Query.Ascending))
+                .Select(doc => doc.Address)
+                .ToImmutableHashSet();
         }
 
         /// <inheritdoc/>
-        public void StageTransactionIds(IDictionary<TxId, bool> txids)
+        public override void StageTransactionIds(IDictionary<TxId, bool> txids)
         {
             StagedTxIds.InsertBulk(
                 txids.Select(kv => new StagedTxIdDoc
@@ -186,13 +174,13 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public void UnstageTransactionIds(ISet<TxId> txids)
+        public override void UnstageTransactionIds(ISet<TxId> txids)
         {
             StagedTxIds.Delete(tx => txids.Contains(tx.TxId));
         }
 
         /// <inheritdoc/>
-        public IEnumerable<TxId> IterateStagedTransactionIds(bool toBroadcast)
+        public override IEnumerable<TxId> IterateStagedTransactionIds(bool toBroadcast)
         {
             IEnumerable<StagedTxIdDoc> docs = StagedTxIds.FindAll();
             if (toBroadcast)
@@ -204,7 +192,7 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public IEnumerable<TxId> IterateTransactionIds()
+        public override IEnumerable<TxId> IterateTransactionIds()
         {
             return _db.FileStorage
                 .Find(TxIdPrefix)
@@ -212,8 +200,7 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public Transaction<T> GetTransaction<T>(TxId txid)
-            where T : IAction, new()
+        public override Transaction<T> GetTransaction<T>(TxId txid)
         {
             LiteFileInfo file = _db.FileStorage.FindById(TxFileId(txid));
             if (file is null)
@@ -223,36 +210,38 @@ namespace Libplanet.Store
 
             using (var stream = new MemoryStream())
             {
-                file.CopyTo(stream);
-                stream.Seek(0, SeekOrigin.Begin);
-                return Transaction<T>.FromBencodex(stream.ToArray());
+                DownloadFile(file, stream);
+
+                var bytes = stream.ToArray();
+                if (bytes.Length != file.Length)
+                {
+                    _logger.Warning(
+                        "The data file for the transaction {TxId} seems corrupted; " +
+                        "it will be treated nonexistent and removed at all.",
+                        txid
+                    );
+                    DeleteTransaction(txid);
+                    return null;
+                }
+
+                return Transaction<T>.FromBencodex(bytes);
             }
         }
 
         /// <inheritdoc/>
-        public void PutTransaction<T>(Transaction<T> tx)
-            where T : IAction, new()
+        public override void PutTransaction<T>(Transaction<T> tx)
         {
-            using (var stream = new MemoryStream())
-            {
-                byte[] encoded = tx.ToBencodex(true);
-                stream.Write(encoded, 0, encoded.Length);
-                stream.Seek(0, SeekOrigin.Begin);
-                _db.FileStorage.Upload(
-                    TxFileId(tx.Id),
-                    tx.Id.ToHex(),
-                    stream);
-            }
+            UploadFile(TxFileId(tx.Id), tx.Id.ToHex(), tx.ToBencodex(true));
         }
 
         /// <inheritdoc/>
-        public bool DeleteTransaction(TxId txid)
+        public override bool DeleteTransaction(TxId txid)
         {
             return _db.FileStorage.Delete(TxFileId(txid));
         }
 
         /// <inheritdoc/>
-        public IEnumerable<HashDigest<SHA256>> IterateBlockHashes()
+        public override IEnumerable<HashDigest<SHA256>> IterateBlockHashes()
         {
             return _db.FileStorage
                 .Find("block/")
@@ -261,77 +250,28 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public Block<T> GetBlock<T>(HashDigest<SHA256> blockHash)
-            where T : IAction, new()
-        {
-            LiteFileInfo file =
-                _db.FileStorage.FindById(BlockFileId(blockHash));
-
-            if (file is null)
-            {
-                return null;
-            }
-
-            using (var stream = new MemoryStream())
-            {
-                file.CopyTo(stream);
-                stream.Seek(0, SeekOrigin.Begin);
-
-                var formatter = new BencodexFormatter<RawBlock>();
-                RawBlock rawBlock = (RawBlock)formatter.Deserialize(stream);
-                HashDigest<SHA256>? previousHash = null;
-
-                if (rawBlock.PreviousHash != null)
-                {
-                    previousHash =
-                        new HashDigest<SHA256>(rawBlock.PreviousHash);
-                }
-
-                return new Block<T>(
-                    index: rawBlock.Index,
-                    difficulty: rawBlock.Difficulty,
-                    nonce: new Nonce(rawBlock.Nonce),
-                    miner: new Address(rawBlock.Miner),
-                    previousHash: previousHash,
-                    timestamp: DateTimeOffset.ParseExact(
-                        rawBlock.Timestamp,
-                        Block<T>.TimestampFormat,
-                        CultureInfo.InvariantCulture
-                    ).ToUniversalTime(),
-                    transactions: GetTransactions<T>(rawBlock.Transactions)
-                );
-            }
-        }
-
-        /// <inheritdoc/>
-        public void PutBlock<T>(Block<T> block)
-            where T : IAction, new()
+        public override void PutBlock<T>(Block<T> block)
         {
             foreach (Transaction<T> tx in block.Transactions)
             {
                 PutTransaction(tx);
             }
 
-            using (var stream = new MemoryStream())
-            {
-                byte[] encoded = block.ToBencodex(true, false);
-                stream.Write(encoded, 0, encoded.Length);
-                stream.Seek(0, SeekOrigin.Begin);
-                _db.FileStorage.Upload(
-                    BlockFileId(block.Hash),
-                    ByteUtil.Hex(block.Hash.ToByteArray()),
-                    stream);
-            }
+            UploadFile(
+                BlockFileId(block.Hash),
+                ByteUtil.Hex(block.Hash.ToByteArray()),
+                block.ToBencodex(true, false)
+            );
         }
 
         /// <inheritdoc/>
-        public bool DeleteBlock(HashDigest<SHA256> blockHash)
+        public override bool DeleteBlock(HashDigest<SHA256> blockHash)
         {
             return _db.FileStorage.Delete(BlockFileId(blockHash));
         }
 
         /// <inheritdoc/>
-        public AddressStateMap GetBlockStates(HashDigest<SHA256> blockHash)
+        public override AddressStateMap GetBlockStates(HashDigest<SHA256> blockHash)
         {
             LiteFileInfo file =
                 _db.FileStorage.FindById(BlockStateFileId(blockHash));
@@ -342,7 +282,7 @@ namespace Libplanet.Store
 
             using (var stream = new MemoryStream())
             {
-                file.CopyTo(stream);
+                DownloadFile(file, stream);
                 stream.Seek(0, SeekOrigin.Begin);
                 var formatter = new BinaryFormatter();
                 return (AddressStateMap)formatter.Deserialize(stream);
@@ -350,7 +290,7 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public void SetBlockStates(
+        public override void SetBlockStates(
             HashDigest<SHA256> blockHash,
             AddressStateMap states)
         {
@@ -362,163 +302,75 @@ namespace Libplanet.Store
                 _db.FileStorage.Upload(
                     BlockStateFileId(blockHash),
                     ByteUtil.Hex(blockHash.ToByteArray()),
-                    stream);
+                    stream
+                );
             }
         }
 
         /// <inheritdoc/>
-        public IEnumerable<Tuple<HashDigest<SHA256>, long>> IterateStateReferences(
+        public override IEnumerable<Tuple<HashDigest<SHA256>, long>> IterateStateReferences(
             string @namespace,
             Address address)
         {
-            var fileId = $"{StateRefIdPrefix}{@namespace}/{address.ToHex()}";
-            LiteFileInfo file = _db.FileStorage.FindById(fileId);
-
-            if (file is null || file.Length == 0)
-            {
-                yield break;
-            }
-
-            int hashSize = HashDigest<SHA256>.Size;
-            int stateReferenceSize = hashSize + sizeof(long);
-            if (file.Length % stateReferenceSize != 0)
-            {
-                throw new FileLoadException(
-                    $"State references file's size ({file.Length}) should be multiple of " +
-                    $"state reference entry size {stateReferenceSize})."
-                );
-            }
-
-            using (var stream = new MemoryStream())
-            {
-                // Note that a stream made by file.OpenRead() does not support
-                // .Seek() operation --- although it implements the interface,
-                // the method throws a NotSupportedException.
-                file.CopyTo(stream);
-
-                var buffer = new byte[stateReferenceSize];
-                long position = stream.Seek(0, SeekOrigin.End);
-
-                for (var i = 1; position - buffer.Length >= 0; i++)
-                {
-                    position = stream.Seek(-buffer.Length * i, SeekOrigin.End);
-                    stream.Read(buffer, 0, buffer.Length);
-                    byte[] hashBytes = buffer.Take(hashSize).ToArray();
-                    long index = BitConverter.ToInt64(buffer, hashSize);
-                    yield return Tuple.Create(
-                        new HashDigest<SHA256>(hashBytes),
-                        index
-                    );
-                }
-            }
+            string collId = StateRefId(@namespace);
+            LiteCollection<StateRefDoc> coll = _db.GetCollection<StateRefDoc>(collId);
+            string addressString = address.ToHex().ToLower();
+            IEnumerable<StateRefDoc> stateRefs = coll.Find(
+                Query.And(
+                    Query.EQ("AddressString", addressString),
+                    Query.All("BlockIndex", Query.Descending)
+                )
+            );
+            return stateRefs
+                .Select(doc => new Tuple<HashDigest<SHA256>, long>(doc.BlockHash, doc.BlockIndex))
+                .OrderByDescending(pair => pair.Item2);
         }
 
         /// <inheritdoc/>
-        public void StoreStateReference<T>(
+        public override void StoreStateReference<T>(
             string @namespace,
             IImmutableSet<Address> addresses,
             Block<T> block)
-            where T : IAction, new()
         {
-            int hashSize = HashDigest<SHA256>.Size;
-            byte[] hashBytes = block.Hash.ToByteArray();
-            byte[] indexBytes = BitConverter.GetBytes(block.Index);
-
-            foreach (Address address in addresses)
-            {
-                string addrHex = address.ToHex();
-                var fileId = $"{StateRefIdPrefix}{@namespace}/{addrHex}";
-                if (!_db.FileStorage.Exists(fileId))
+            string collId = StateRefId(@namespace);
+            LiteCollection<StateRefDoc> coll = _db.GetCollection<StateRefDoc>(collId);
+            coll.InsertBulk(
+                addresses.Select(addr => new StateRefDoc
                 {
-                    _db.FileStorage.Upload(
-                        fileId,
-                        addrHex,
-                        new MemoryStream());
-                }
-
-                LiteFileInfo file = _db.FileStorage.FindById(fileId);
-                using (var temp = new MemoryStream())
-                {
-                    file.CopyTo(temp);
-                    temp.Seek(0, SeekOrigin.Begin);
-                    byte[] prev = temp.ToArray();
-
-                    using (LiteFileStream stream = file.OpenWrite())
-                    {
-                        stream.Write(prev, 0, prev.Length);
-                        stream.Write(hashBytes, 0, hashSize);
-                        stream.Write(indexBytes, 0, sizeof(long));
-                    }
-                }
-            }
+                    Address = addr,
+                    BlockIndex = block.Index,
+                    BlockHash = block.Hash,
+                })
+            );
+            coll.EnsureIndex("AddressString");
+            coll.EnsureIndex("BlockIndex");
         }
 
         /// <inheritdoc/>
-        public void ForkStateReferences<T>(
+        public override void ForkStateReferences<T>(
             string srcNamespace,
             string destNamespace,
             Block<T> branchPoint,
             IImmutableSet<Address> addressesToStrip)
-            where T : IAction, new()
         {
-            long branchPointIndex = branchPoint.Index;
-            List<LiteFileInfo> files =
-                _db.FileStorage
-                    .Find($"{StateRefIdPrefix}{srcNamespace}")
-                    .ToList();
+            string srcCollId = StateRefId(srcNamespace);
+            string dstCollId = StateRefId(destNamespace);
+            LiteCollection<StateRefDoc> srcColl = _db.GetCollection<StateRefDoc>(srcCollId),
+                                        dstColl = _db.GetCollection<StateRefDoc>(dstCollId);
 
-            if (!files.Any() && addressesToStrip.Any())
+            dstColl.InsertBulk(srcColl.Find(Query.LTE("BlockIndex", branchPoint.Index)));
+
+            if (dstColl.Count() < 1 && addressesToStrip.Any())
             {
                 throw new NamespaceNotFoundException(
                     srcNamespace,
-                    "The source namespace to be forked does not exist.");
-            }
-
-            foreach (LiteFileInfo srcFile in files)
-            {
-                string destId =
-                    $"{StateRefIdPrefix}{destNamespace}/{srcFile.Filename}";
-                _db.FileStorage.Upload(
-                    destId,
-                    srcFile.Filename,
-                    new MemoryStream());
-
-                LiteFileInfo destFile = _db.FileStorage.FindById(destId);
-                using (LiteFileStream srcStream = srcFile.OpenRead())
-                using (LiteFileStream destStream = destFile.OpenWrite())
-                {
-                    while (srcStream.Position < srcStream.Length)
-                    {
-                        var hashBytes = new byte[HashDigest<SHA256>.Size];
-                        var indexBytes = new byte[sizeof(long)];
-
-                        srcStream.Read(hashBytes, 0, hashBytes.Length);
-                        srcStream.Read(indexBytes, 0, indexBytes.Length);
-
-                        long currentIndex =
-                            BitConverter.ToInt64(indexBytes, 0);
-
-                        if (currentIndex <= branchPointIndex)
-                        {
-                            destStream.Write(hashBytes, 0, hashBytes.Length);
-                            destStream.Write(indexBytes, 0, indexBytes.Length);
-                        }
-                        else
-                        {
-                            break;
-                        }
-                    }
-                }
-
-                if (destFile.Length == 0)
-                {
-                    _db.FileStorage.Delete(destId);
-                }
+                    "The source namespace to be forked does not exist."
+                );
             }
         }
 
         /// <inheritdoc/>
-        public IEnumerable<KeyValuePair<Address, long>> ListTxNonces(string @namespace)
+        public override IEnumerable<KeyValuePair<Address, long>> ListTxNonces(string @namespace)
         {
             var collectionId = $"{NonceIdPrefix}{@namespace}";
             LiteCollection<BsonDocument> collection = _db.GetCollection<BsonDocument>(collectionId);
@@ -539,7 +391,7 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public long GetTxNonce(string @namespace, Address address)
+        public override long GetTxNonce(string @namespace, Address address)
         {
             var collectionId = $"{NonceIdPrefix}{@namespace}";
             LiteCollection<BsonDocument> collection = _db.GetCollection<BsonDocument>(collectionId);
@@ -549,7 +401,7 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public void IncreaseTxNonce(string @namespace, Address signer, long delta = 1)
+        public override void IncreaseTxNonce(string @namespace, Address signer, long delta = 1)
         {
             long nextNonce = GetTxNonce(@namespace, signer) + delta;
             var collectionId = $"{NonceIdPrefix}{@namespace}";
@@ -559,13 +411,13 @@ namespace Libplanet.Store
         }
 
         /// <inheritdoc/>
-        public long CountTransactions()
+        public override long CountTransactions()
         {
             return _db.FileStorage.Find(TxIdPrefix).Count();
         }
 
         /// <inheritdoc/>
-        public long CountBlocks()
+        public override long CountBlocks()
         {
             return _db.FileStorage.Find("block/").Count();
         }
@@ -573,6 +425,25 @@ namespace Libplanet.Store
         public void Dispose()
         {
             _db?.Dispose();
+        }
+
+        internal override RawBlock? GetRawBlock(HashDigest<SHA256> blockHash)
+        {
+            LiteFileInfo file =
+                _db.FileStorage.FindById(BlockFileId(blockHash));
+            if (file is null)
+            {
+                return null;
+            }
+
+            using (var stream = new MemoryStream())
+            {
+                DownloadFile(file, stream);
+                stream.Seek(0, SeekOrigin.Begin);
+
+                var formatter = new BencodexFormatter<RawBlock>();
+                return (RawBlock)formatter.Deserialize(stream);
+            }
         }
 
         private string TxFileId(TxId txid)
@@ -590,6 +461,13 @@ namespace Libplanet.Store
             return $"state/{blockHash}";
         }
 
+        private string StateRefId(string @namespace)
+        {
+            string ns =
+                ByteUtil.Hex(namespaceHasher.ComputeHash(Encoding.UTF8.GetBytes(@namespace)));
+            return $"{StateRefIdPrefix}{ns}";
+        }
+
         private IEnumerable<Transaction<T>> GetTransactions<T>(
             IEnumerable transactions
         )
@@ -601,9 +479,68 @@ namespace Libplanet.Store
                 .Where(tx => tx != null);
         }
 
+        private void UploadFile(string fileId, string filename, byte[] bytes)
+        {
+            using (var stream = new MemoryStream(bytes))
+            {
+                stream.Seek(0, SeekOrigin.Begin);
+                _db.FileStorage.Upload(fileId, filename, stream);
+            }
+        }
+
+        private void DownloadFile(LiteFileInfo file, Stream stream)
+        {
+            file.CopyTo(stream);
+
+            if (stream.Length > file.Length)
+            {
+                stream.SetLength(file.Length);
+            }
+        }
+
         private LiteCollection<HashDoc> IndexCollection(string @namespace)
         {
             return _db.GetCollection<HashDoc>($"{IndexColPrefix}{@namespace}");
+        }
+
+        internal class StateRefDoc
+        {
+            public string AddressString { get; set; }
+
+            public long BlockIndex { get; set; }
+
+            public string BlockHashString { get; set; }
+
+            [BsonId]
+            public string Id => AddressString + BlockHashString;
+
+            [BsonIgnore]
+            public Address Address
+            {
+                get
+                {
+                    return new Address(AddressString);
+                }
+
+                set
+                {
+                    AddressString = value.ToHex().ToLower();
+                }
+            }
+
+            [BsonIgnore]
+            public HashDigest<SHA256> BlockHash
+            {
+                get
+                {
+                    return HashDigest<SHA256>.FromString(BlockHashString);
+                }
+
+                set
+                {
+                    BlockHashString = value.ToString();
+                }
+            }
         }
 
         private class HashDoc